requirements.txt
flask>=2.0.1
neo4j>=5.27.0
numpy>=1.26.4
torch>=2.5.1
torchvision>=0.20.1
pillow>=10.2.0
matplotlib>=3.10.0
pandas>=2.2.0
scikit-learn>=1.4.0
opencv-python>=4.10.0
python-dotenv>=1.0.0
werkzeug>=3.0.1
pycocotools>=2.0.7
tqdm>=4.66.2
requests>=2.31.0
